requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.1
deep-translator==1.11.4
pymongo==4.7.1
pytz==2024.1
Flask==3.0.3
psutil==5.9.8
telegraph[aio]==2.2.0
gtts==2.5.4
yt-dlp==2025.1.12
qrcode[pil]==8.0
aiohttp[speedups]
requests
TgCrypto
aiogram